Logging: Update log messages to include more info
Update log messages to include additional information for diagnosing
problems (e.g., buffer sizes, file types, file ownership).

Additionally, change the printf() conversion specifier for uid_t/gid_t
from unsigned int to unsigned long.

Tested:
- AlmaLinux 9.5, 8.10
- Arch Linux
- CentOS Linux Stream 9, 8.5.2111, 7.9.2009, 6.10
- Debian sid, 12.9, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 41, 40
- FreeBSD 14.2, 13.4
- NetBSD 10.1, 9.4
- OpenBSD 7.6, 7.5
- openSUSE 15.6, 15.5
- Ubuntu 24.10, 24.04.2, 22.04.5, 20.04.6, 18.04.6, 16.04.7, 14.04.6

+++ src/munged/gids.c
@@ -608,14 +608,14 @@ _gids_gid_add (hash_t gid_hash, uid_t uid, gid_t gid)
 
     if (!(g = hash_find (gid_hash, &uid))) {
         if (!(g = _gids_gid_head_create (uid))) {
-            log_msg (LOG_WARNING, "Failed to allocate gid head for uid=%u",
-                    (unsigned int) uid);
+            log_msg (LOG_WARNING, "Failed to allocate gid head for uid=%lu",
+                    (unsigned long) uid);
             return (-1);
         }
         if (!hash_insert (gid_hash, &g->uid, g)) {
             log_msg (LOG_WARNING,
-                    "Failed to insert gid head for uid=%u into gid hash",
-                    (unsigned int) uid);
+                    "Failed to insert gid head for uid=%lu into gid hash",
+                    (unsigned long) uid);
             _gids_gid_head_destroy (g);
             return (-1);
         }
@@ -630,8 +630,9 @@ _gids_gid_add (hash_t gid_hash, uid_t uid, gid_t gid)
         return (0);
     }
     if (!(node = _gids_gid_node_create (gid))) {
-        log_msg (LOG_WARNING, "Failed to allocate gid node for uid=%u gid=%u",
-                (unsigned int) uid, (unsigned int) gid);
+        log_msg (LOG_WARNING,
+                "Failed to allocate gid node for uid=%lu gid=%lu",
+                (unsigned long) uid, (unsigned long) gid);
         return (-1);
     }
     node->next = *nodep;
@@ -650,13 +651,13 @@ _gids_uid_add (hash_t uid_hash, const char *user, uid_t uid)
     uid_node_p u;
 
     if (!(u = _gids_uid_node_create (user, uid))) {
-        log_msg (LOG_WARNING, "Failed to allocate uid node for \"%s\" uid=%u",
-                user, (unsigned int) uid);
+        log_msg (LOG_WARNING, "Failed to allocate uid node for \"%s\" uid=%lu",
+                user, (unsigned long) uid);
     }
     else if (!hash_insert (uid_hash, u->user, u)) {
         log_msg (LOG_WARNING,
-                "Failed to insert uid node for \"%s\" uid=%u into uid hash",
-                user, (unsigned int) uid);
+                "Failed to insert uid node for \"%s\" uid=%lu into uid hash",
+                user, (unsigned long) uid);
         _gids_uid_node_destroy (u);
     }
     else {
